Implementations...

+++ src/main/java/app/repository/AccountRepository.java
@@ -2,5 +2,5 @@ package app.repository;
 
 import app.model.Account;
 
-public interface AccountRepository extends GenericRepository<Account, Long> {
+public interface AccountRepository extends GenericRepository<Account, Integer> {
 }
